requirements.txt
django-tree-queries>=0.11.0
djangorestframework>=3.13
djangoajax>=3.3
django-markitup>=4.0.0
markdown>=3.4.1
Pillow
django>=4.0
django_filter>=22.1
djgentelella
